+++ osu.Game.Tournament/TournamentGameBase.cs
@@ -103,7 +103,7 @@ namespace osu.Game.Tournament
                         },
                         new OsuSpriteText
                         {
-                            Text = "窗口太窄了,拉宽一些",
+                            Text = "窗口太窄了，拉宽一些",
                             Font = OsuFont.Default.With(weight: "bold"),
                             Colour = Color4.White,
                             Padding = new MarginPadding(20)
@@ -121,10 +121,9 @@ namespace osu.Game.Tournament
                 using (var sr = new StreamReader(stream))
                     ladder = JsonConvert.DeserializeObject<LadderInfo>(sr.ReadToEnd());
             }
-            else
-            {
+
+            if (ladder == null)
                 ladder = new LadderInfo();
-            }
 
             if (ladder.Ruleset.Value == null)
                 ladder.Ruleset.Value = RulesetStore.AvailableRulesets.First();
@@ -202,9 +201,11 @@ namespace osu.Game.Tournament
             {
                 foreach (var p in t.Players)
                 {
-                    if (p.Username == null || p.Statistics == null)
+                    if (string.IsNullOrEmpty(p.Username) || p.Statistics == null)
+                    {
                         PopulateUser(p);
-                    addedInfo = true;
+                        addedInfo = true;
+                    }
                 }
             }
 
@@ -237,6 +238,24 @@ namespace osu.Game.Tournament
                     if (b.BeatmapInfo == null)
                         // if online population couldn't be performed, ensure we don't leave a null value behind
                         r.Beatmaps.Remove(b);
+                }
+            }
+
+            foreach (var t in ladder.Teams)
+            {
+                foreach (var s in t.SeedingResults)
+                {
+                    foreach (var b in s.Beatmaps)
+                    {
+                        if (b.BeatmapInfo == null && b.ID > 0)
+                        {
+                            var req = new GetBeatmapRequest(new BeatmapInfo { OnlineBeatmapID = b.ID });
+                            req.Perform(API);
+                            b.BeatmapInfo = req.Result?.ToBeatmap(RulesetStore);
+
+                            addedInfo = true;
+                        }
+                    }
                 }
             }