+++ Changelog.md
@@ -574,3 +574,4 @@
     *   [m] [Fix nullref when deleting teams in the ladder screen #8123]
     *   [m] [因为和现有issue模板冲突，未整合] [Comment out log locations in bug report template #8121]
     *   [m] [Add torus font #8104]
+    *   [m] [**Add tournament seeding screen #8105**]